Move base_test_channel to Core
