+++ Tribler/Test/Core/Modules/RestApi/base_api_test.py
@@ -13,7 +13,7 @@ from twisted.web.iweb import IBodyProducer
 from Tribler.Core.Utilities.network_utils import get_random_port
 from Tribler.Core.Utilities.twisted_thread import reactor
 from Tribler.Core.version import version_id
-from Tribler.Test.Core.Modules.Channel.base_test_channel import BaseTestChannel
+from Tribler.Test.Core.base_test_channel import BaseTestChannel
 from Tribler.Test.test_as_server import TestAsServer
 from Tribler.dispersy.util import blocking_call_on_reactor_thread
