+++ Tribler/Test/Core/base_test_channel.py
